精简 Demo
