+++ app/src/main/res/layout/flap_item_simple_text.xml
@@ -10,7 +10,8 @@
     <TextView
         android:id="@+id/tv_content"
         android:layout_width="match_parent"
-        android:layout_height="50dp"
+        android:layout_height="wrap_content"
+        android:minHeight="50dp"
         android:gravity="center"
         android:paddingLeft="16dp"
         android:text="content"
